Update README.md
# Greenhouse-Gas-Analysis
The aim of this project is to determine the rate at which the release of greenhouse gases into the atmosphere causes climate change and how much it increases every year. To determine if the amount release differs in regions.
Mocrosoft SQL was used for data extraction.

Power Bi was used for the analysis ans dashboard creation.

The ininally 50 columns and 1158 rows dataset was stramlined to 16 columns and 514 rows. This is becasue the unimportant columns to the analysis was deleted and some rows that made no meaning were also removed.

Thank you for stopping by

The story can be found https://medium.com/@lilian.c.nwafor/greenhouse-gas-ghg-air-emissions-an-analysis-of-annual-estimates-of-greenhouse-gas-emissions-7959fa3c2bd0